src/connect_azure_sql.py
"""
This module is used to connect to an Azure SQL database using the pyodbc module. 

It establishes a connection using a predefined connection string, creates a cursor object, 
executes a SQL query to fetch all records from a specified table, and then prints the results.

Usage:
    python connect_azure_sql.py
"""

# Import the necessary modules
import os
import logging
import pyodbc
from pyodbc import Error

# Set up logging
logging.basicConfig(level=logging.INFO)

# Get the connection string from environment variables
CONNECTION_STRING = os.getenv("CONNECTION_STRING")


def connect_to_database(connection_string):
    """
    Connect to an Azure SQL database.

    Parameters:
    connection_string (str): The connection string for the Azure SQL database.

    Returns:
    pyodbc.Cursor: The cursor object for the database connection.
    """
    try:
        # Connect to the database
        connection = pyodbc.connect(connection_string)

        # Create a cursor object
        cursor = connection.cursor()

        return cursor

    except pyodbc.Error as e:
        logging.error(
            f"An error occurred while connecting to the database: {str(e)}")
        raise  # re-raise the exception


def main():
    """
    The main function that connects to the database and performs the necessary operations.
    """
    # Connect to the database
    cursor = connect_to_database(CONNECTION_STRING)

    if cursor is not None:
        # Add the code to execute the SQL query and print the results here
        pass


if __name__ == "__main__":
    main()
